fix(static-helpers): unique stable-hash tag for MapGroupBy + extract groupBy to groupby.rs (under 2000-line cap)
