+++ crates/perry-hir/src/stable_hash/expr.rs
@@ -497,7 +497,7 @@ impl SH for Expr {
             Expr::ObjectValues(e) => { tag(h, 392); e.as_ref().hash(h); }
             Expr::ObjectEntries(e) => { tag(h, 393); e.as_ref().hash(h); }
             Expr::ObjectGroupBy { items, key_fn } => { tag(h, 394); items.as_ref().hash(h); key_fn.as_ref().hash(h); }
-            Expr::MapGroupBy { items, key_fn } => { tag(h, 12042); items.as_ref().hash(h); key_fn.as_ref().hash(h); }
+            Expr::MapGroupBy { items, key_fn } => { tag(h, 12044); items.as_ref().hash(h); key_fn.as_ref().hash(h); }
             Expr::ObjectRest { object, exclude_keys, } => { tag(h, 395); object.as_ref().hash(h); exclude_keys.hash(h); }
             Expr::ArrayIsArray(e) => { tag(h, 396); e.as_ref().hash(h); }
             Expr::ArrayFrom(e) => { tag(h, 397); e.as_ref().hash(h); }
