+++ crates/perry-runtime/src/object/mod.rs
@@ -34,6 +34,7 @@ mod delete_rest;
 mod field_get_set;
 mod field_set_by_name;
 mod global_this;
+mod groupby;
 mod has_own_helpers;
 mod instanceof;
 mod native_call_method;
@@ -60,6 +61,7 @@ pub use delete_rest::*;
 pub use field_get_set::*;
 pub use field_set_by_name::*;
 pub use global_this::*;
+pub use groupby::*;
 pub use instanceof::*;
 pub use native_call_method::*;
 pub use native_module::*;
